refactor: moved technology-card-component to technology-viewer/components/
