+++ frontend/src/app/modules/admin/components/technology-dashboard/technology-dashboard.component.ts
@@ -7,7 +7,7 @@ import Technology from '../../../../shared/interfaces/Technology';
 import { TechnologyViewerComponent } from '../../../../shared/components/technology-viewer/technology-viewer.component';
 import { TechnologyService } from '../../services/technology/technology.service';
 import { ScreenComponent } from '../../../../shared/components/screen/screen.component';
-import { TechnologyCardComponent } from '../../../../shared/components/technology-card/technology-card.component';
+import { TechnologyCardComponent } from '../../../../shared/components/technology-viewer/components/technology-card/technology-card.component';
 import { AlertComponent } from '../../../../shared/components/alert/alert.component';
 
 @Component({
