+++ frontend/src/app/shared/components/technology-viewer/components/technology-card/technology-card.component.html
